feat: improve nip21 prefix handlers
